requirements.txt
Flask
Flask-CORS
Flask-Migrate
Flask-Script
Flask-SQLAlchemy
marshmallow
passlib
PyYAML
sqlalchemy-utils
